+++ ICD.Connect.Audio/Controls/AbstractVolumeLevelDeviceControl.cs
@@ -1,7 +1,12 @@
 ﻿using System;
+using System.Collections.Generic;
 using ICD.Common.Properties;
 using ICD.Common.Utils;
 using ICD.Common.Utils.Extensions;
+using ICD.Connect.API.Commands;
+using ICD.Connect.API.Nodes;
+using ICD.Connect.Audio.Console;
+using ICD.Connect.Audio.EventArguments;
 using ICD.Connect.Audio.Repeaters;
 using ICD.Connect.Devices;
 using ICD.Connect.Devices.Controls;
@@ -165,6 +170,16 @@ namespace ICD.Connect.Audio.Controls
 
 		#endregion
 
+		/// <summary>
+		/// Constructor
+		/// </summary>
+		/// <param name="parent">Device this control belongs to</param>
+		/// <param name="id">Id of this control in the device</param>
+		protected AbstractVolumeLevelDeviceControl(T parent, int id) : base(parent, id)
+		{
+			m_RepeaterCriticalSection = new SafeCriticalSection();
+		}
+
 		#region Abstract Methods
 
 		public abstract void SetVolumeRaw(float volume);
@@ -299,14 +314,63 @@ namespace ICD.Connect.Audio.Controls
 
 		#endregion
 
+		#region Console
+
 		/// <summary>
-		/// Constructor
+		/// Gets the child console nodes.
 		/// </summary>
-		/// <param name="parent">Device this control belongs to</param>
-		/// <param name="id">Id of this control in the device</param>
-		protected AbstractVolumeLevelDeviceControl(T parent, int id) : base(parent, id)
+		/// <returns></returns>
+		public override IEnumerable<IConsoleNodeBase> GetConsoleNodes()
 		{
-			m_RepeaterCriticalSection = new SafeCriticalSection();
+			foreach (IConsoleNodeBase node in GetBaseConsoleNodes())
+				yield return node;
+
+			foreach (IConsoleNodeBase node in VolumeLevelDeviceControlConsole.GetConsoleNodes(this))
+				yield return node;
 		}
+
+		/// <summary>
+		/// Wrokaround for "unverifiable code" warning.
+		/// </summary>
+		/// <returns></returns>
+		private IEnumerable<IConsoleNodeBase> GetBaseConsoleNodes()
+		{
+			return base.GetConsoleNodes();
+		}
+
+		/// <summary>
+		/// Calls the delegate for each console status item.
+		/// </summary>
+		/// <param name="addRow"></param>
+		public override void BuildConsoleStatus(AddStatusRowDelegate addRow)
+		{
+			base.BuildConsoleStatus(addRow);
+
+			VolumeLevelDeviceControlConsole.BuildConsoleStatus(this, addRow);
+		}
+
+		/// <summary>
+		/// Gets the child console commands.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<IConsoleCommand> GetConsoleCommands()
+		{
+			foreach (IConsoleCommand command in GetBaseConsoleCommands())
+				yield return command;
+
+			foreach (IConsoleCommand command in VolumeLevelDeviceControlConsole.GetConsoleCommands(this))
+				yield return command;
+		}
+
+		/// <summary>
+		/// Workaround for "unverifiable code" warning.
+		/// </summary>
+		/// <returns></returns>
+		private IEnumerable<IConsoleCommand> GetBaseConsoleCommands()
+		{
+			return base.GetConsoleCommands();
+		}
+
+		#endregion
 	}
 }