+++ ICD.Connect.Audio/Controls/IVolumeLevelDeviceControl.cs
@@ -1,6 +1,8 @@
 ﻿using System;
-using ICD.Common.Properties;
 using ICD.Common.Utils;
+using ICD.Connect.API.Attributes;
+using ICD.Connect.Audio.EventArguments;
+using ICD.Connect.Audio.Proxies.Controls;
 
 namespace ICD.Connect.Audio.Controls
 {
@@ -13,7 +15,7 @@ namespace ICD.Connect.Audio.Controls
 		/// <summary>
 		/// Raised when the raw volume changes.
 		/// </summary>
-		[PublicAPI]
+		[ApiEvent(VolumeLevelDeviceControlApi.EVENT_VOLUME_CHANGED, VolumeLevelDeviceControlApi.HELP_EVENT_VOLUME_CHANGED)]
 		event EventHandler<VolumeDeviceVolumeChangedEventArgs> OnVolumeChanged;
 
 		#region Properties
@@ -21,33 +23,37 @@ namespace ICD.Connect.Audio.Controls
 		/// <summary>
 		/// Gets the current volume, in the parent device's format
 		/// </summary>
-		[PublicAPI]
+		[ApiProperty(VolumeLevelDeviceControlApi.PROPERTY_VOLUME_RAW, VolumeLevelDeviceControlApi.HELP_PROPERTY_VOLUME_RAW)]
 		float VolumeRaw { get; }
 
 		/// <summary>
 		/// Gets the current volume positon, 0 - 1
 		/// </summary>
-		[PublicAPI]
+		[ApiProperty(VolumeLevelDeviceControlApi.PROPERTY_VOLUME_POSITION,
+			VolumeLevelDeviceControlApi.HELP_PROPERTY_VOLUME_POSITION)]
 		float VolumePosition { get; }
 
 		/// <summary>
 		/// Gets the current volume, in string representation
 		/// </summary>
-		[PublicAPI]
+		[ApiProperty(VolumeLevelDeviceControlApi.PROPERTY_VOLUME_STRING,
+			VolumeLevelDeviceControlApi.HELP_PROPERTY_VOLUME_STRING)]
 		string VolumeString { get; }
 
 		/// <summary>
 		/// Maximum value for the raw volume level
 		/// This could be the maximum permitted by the device/control, or a safety max
 		/// </summary>
-		[PublicAPI]
+		[ApiProperty(VolumeLevelDeviceControlApi.PROPERTY_VOLUME_RAW_MAX,
+			VolumeLevelDeviceControlApi.HELP_PROPERTY_VOLUME_RAW_MAX)]
 		float? VolumeRawMax { get; }
 
 		/// <summary>
 		/// Minimum value for the raw volume level
 		/// This could be the minimum permitted by the device/control, or a safety min
 		/// </summary>
-		[PublicAPI]
+		[ApiProperty(VolumeLevelDeviceControlApi.PROPERTY_VOLUME_RAW_MIN,
+			VolumeLevelDeviceControlApi.HELP_PROPERTY_VOLUME_RAW_MIN)]
 		float? VolumeRawMin { get; }
 
 		#endregion
@@ -58,26 +64,29 @@ namespace ICD.Connect.Audio.Controls
 		/// Sets the raw volume. This will be clamped to the min/max and safety min/max.
 		/// </summary>
 		/// <param name="volume"></param>
-		[PublicAPI]
+		[ApiMethod(VolumeLevelDeviceControlApi.METHOD_SET_VOLUME_RAW, VolumeLevelDeviceControlApi.HELP_METHOD_SET_VOLUME_RAW)]
 		void SetVolumeRaw(float volume);
 
 		/// <summary>
 		/// Sets the volume position, from 0-1
 		/// </summary>
 		/// <param name="position"></param>
-		[PublicAPI]
+		[ApiMethod(VolumeLevelDeviceControlApi.METHOD_SET_VOLUME_POSITION,
+			VolumeLevelDeviceControlApi.HELP_METHOD_SET_VOLUME_POSITION)]
 		void SetVolumePosition(float position);
 
 		/// <summary>
 		/// Increments the volume once.
 		/// </summary>
-		[PublicAPI]
+		[ApiMethod(VolumeLevelDeviceControlApi.METHOD_VOLUME_LEVEL_INCREMENT,
+			VolumeLevelDeviceControlApi.HELP_METHOD_VOLUME_LEVEL_INCREMENT)]
 		void VolumeLevelIncrement(float incrementValue);
 
 		/// <summary>
 		/// Decrements the volume once.
 		/// </summary>
-		[PublicAPI]
+		[ApiMethod(VolumeLevelDeviceControlApi.METHOD_VOLUME_LEVEL_DECREMENT,
+			VolumeLevelDeviceControlApi.HELP_METHOD_VOLUME_LEVEL_DECREMENT)]
 		void VolumeLevelDecrement(float decrementValue);
 
 		#endregion