+++ ICD.Connect.Audio/Proxies/Controls/VolumeMuteDeviceControlApi.cs
@@ -1,0 +1,9 @@
+﻿namespace ICD.Connect.Audio.Proxies.Controls
+{
+	public static class VolumeMuteDeviceControlApi
+	{
+		public const string METHOD_SET_VOLUME_MUTE = "SetVolumeMute";
+
+		public const string HELP_METHOD_SET_VOLUME_MUTE = "Sets the mute state.";
+	}
+}